concert_ticketing/requirements.txt
django-cms==3.11.0
djangorestframework==3.14.0
gspread==5.8.0
oauth2client==4.1.3
celery==5.3.1
redis==4.5.5  # For Celery broker
django-cors-headers==3.14.0
django-celery-results==3.3.1
gspread==6.0.0
google-auth==2.28.1
celery==5.3.6
redis==5.0.1